+++ evaluate/pii-masking-300k/read_token_template.py
