+++ requirements/Dockerfile
@@ -1,7 +1,7 @@
 FROM ollama/ollama
 # Sleep to wait for ollama server to be ready
 RUN ollama serve & sleep 5 && ollama pull llama3.2 
-RUN apt-get update && apt-get install python3 python3-pip -y && apt-get clean autoclean && rm -rf /var/lib/apt/lists/*
+RUN apt-get update && apt-get install -y dos2unix && apt-get install python3 python3-pip -y && apt-get clean autoclean && rm -rf /var/lib/apt/lists/*
 RUN --mount=type=bind,source=requirements.txt,target=requirements.txt pip3 install -r requirements.txt
 WORKDIR /entry
 COPY entrypoint.sh .
